+++ src/yaplox/stmt.py
@@ -20,6 +20,10 @@ class StmtVisitor(ABC):
         raise NotImplementedError
 
     @abstractmethod
+    def visit_function_stmt(self, stmt: "Function"):
+        raise NotImplementedError
+
+    @abstractmethod
     def visit_if_stmt(self, stmt: "If"):
         raise NotImplementedError
 
@@ -58,6 +62,17 @@ class Expression(Stmt):
     def accept(self, visitor: StmtVisitor):
         """ Create a accept method that calls the visitor. """
         return visitor.visit_expression_stmt(self)
+
+
+class Function(Stmt):
+    def __init__(self, name: Token, params: List[Token], body: List[Stmt]):
+        self.name = name
+        self.params = params
+        self.body = body
+
+    def accept(self, visitor: StmtVisitor):
+        """ Create a accept method that calls the visitor. """
+        return visitor.visit_function_stmt(self)
 
 
 class If(Stmt):